ERC: Track: Rewrite and rename `erc-track-find-face'
For clarification purposes.  No functional changes.

* lisp/erc/erc-track.el (erc-track-find-face): Rename it to
`erc-track-select-face'.  Rewrite it so that it is very clear what is
the current algorithm, changing the parameters it takes.  No
functional changes. Performance improvements.  Clarify the
documentation and remove the part on some faces being lists, which
clearly doesn't apply.

* lisp/erc/erc-track.el (erc-track-modified-channels): Replace calls
to `erc-track-find-face' by calls to `erc-track-select-face',
preserving the existing behavior.

* lisp/erc/erc-track.el (erc-modified-channels-alist): Change the
reference to `erc-track-select-face' in the documentation following
the rename.